Rename Chat.ts to ChatManager.ts and fix up the state management for ChatView.ts
